# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Help Desk API","description":"AI-powered Help Desk system with PostgreSQL","version":"2.0.0"},"paths":{"/auth/register":{"post":{"tags":["authentication"],"summary":"Register User","description":"Регистрация нового пользователя","operationId":"register_user_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["authentication"],"summary":"Login User","description":"Вход пользователя (получение токена)","operationId":"login_user_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login/form":{"post":{"tags":["authentication"],"summary":"Login User Form","description":"Вход через OAuth2 форму (для Swagger UI)","operationId":"login_user_form_auth_login_form_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_user_form_auth_login_form_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["authentication"],"summary":"Get Current User","description":"Получение информации о текущем пользователе","operationId":"get_current_user_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/logout":{"post":{"tags":["authentication"],"summary":"Logout User","description":"Выход пользователя (удаление токена)","operationId":"logout_user_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/tickets/create":{"post":{"tags":["tickets"],"summary":"Create Ticket","description":"Создает новый тикет и автоматически обрабатывает его с помощью ИИ","operationId":"create_ticket_tickets_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tickets/{ticket_id}":{"get":{"tags":["tickets"],"summary":"Get Ticket","description":"Получает тикет по ID","operationId":"get_ticket_tickets__ticket_id__get","parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["tickets"],"summary":"Update Ticket","description":"Обновляет тикет","operationId":"update_ticket_tickets__ticket_id__put","parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Ticket Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["tickets"],"summary":"Delete Ticket","description":"Удаляет тикет (soft delete - меняет статус на closed)","operationId":"delete_ticket_tickets__ticket_id__delete","parameters":[{"name":"ticket_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Ticket Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tickets":{"get":{"tags":["tickets"],"summary":"List Tickets","description":"Список тикетов с фильтрацией","operationId":"list_tickets_tickets_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TicketStatus"},{"type":"null"}],"title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketResponse"},"title":"Response List Tickets Tickets Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_user_form_auth_login_form_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_user_form_auth_login_form_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IssueType":{"type":"string","enum":["simple","complex","auto_resolvable"],"title":"IssueType","description":"Типы проблем"},"TicketCreate":{"properties":{"source":{"$ref":"#/components/schemas/TicketSource"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"subject":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject"},"body":{"type":"string","title":"Body"},"language":{"$ref":"#/components/schemas/TicketLanguage","default":"ru"}},"type":"object","required":["source","user_id","body"],"title":"TicketCreate","description":"Схема для создания тикета"},"TicketLanguage":{"type":"string","enum":["ru","kk","en"],"title":"TicketLanguage","description":"Языки"},"TicketPriority":{"type":"string","enum":["low","medium","high","critical"],"title":"TicketPriority","description":"Приоритеты"},"TicketResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"source":{"$ref":"#/components/schemas/TicketSource"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"subject":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject"},"body":{"type":"string","title":"Body"},"language":{"$ref":"#/components/schemas/TicketLanguage"},"category_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Category Id"},"priority":{"anyOf":[{"$ref":"#/components/schemas/TicketPriority"},{"type":"null"}]},"issue_type":{"anyOf":[{"$ref":"#/components/schemas/IssueType"},{"type":"null"}]},"ai_confidence":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Ai Confidence"},"assigned_department_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Assigned Department Id"},"assigned_operator_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Assigned Operator Id"},"status":{"$ref":"#/components/schemas/TicketStatus"},"auto_resolved":{"type":"boolean","title":"Auto Resolved"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"closed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Closed At"}},"type":"object","required":["id","source","user_id","subject","body","language","category_id","priority","issue_type","ai_confidence","assigned_department_id","assigned_operator_id","status","auto_resolved","created_at","updated_at","closed_at"],"title":"TicketResponse","description":"Схема ответа с тикетом"},"TicketSource":{"type":"string","enum":["email","chat","portal","phone"],"title":"TicketSource","description":"Источники обращения"},"TicketStatus":{"type":"string","enum":["new","auto_resolved","in_work","waiting","closed"],"title":"TicketStatus","description":"Статусы тикетов"},"TicketUpdate":{"properties":{"status":{"anyOf":[{"$ref":"#/components/schemas/TicketStatus"},{"type":"null"}]},"priority":{"anyOf":[{"$ref":"#/components/schemas/TicketPriority"},{"type":"null"}]},"category_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Category Id"},"assigned_department_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Assigned Department Id"},"assigned_operator_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Assigned Operator Id"}},"type":"object","title":"TicketUpdate","description":"Схема для обновления тикета"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"}},"type":"object","required":["access_token","user_id","email","name","role"],"title":"TokenResponse","description":"Схема ответа с токеном"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin","description":"Схема для входа пользователя"},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"name":{"type":"string","title":"Name"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"position":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Position"},"role":{"type":"string","title":"Role","default":"client"}},"type":"object","required":["email","password","name"],"title":"UserRegister","description":"Схема для регистрации пользователя"},"UserResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"name":{"type":"string","title":"Name"},"position":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Position"},"role":{"type":"string","title":"Role"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","email","phone","name","position","role","created_at"],"title":"UserResponse","description":"Схема ответа с пользователем"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/login"}}}}}}